update: 2.0.0

+++ mm-service/src/main/java/cn/mmanager/service/MQTT/impl/NmqsServiceImpl.java
@@ -1,10 +1,15 @@
 package cn.mmanager.service.MQTT.impl;
 
+import cn.mmanager.dao.MQTT.MqPageMapper;
 import cn.mmanager.dao.MQTT.NmqsMapper;
+import cn.mmanager.model.dto.MqPageDto;
 import cn.mmanager.model.pojo.MQPage;
 import cn.mmanager.model.pojo.NmqsToken;
 import cn.mmanager.service.MQTT.MqPageService;
 import cn.mmanager.service.MQTT.NmqsService;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,19 +20,11 @@ import java.util.Map;
  * @author NicholasLD
  * @createTime 2023/4/13 14:36
  */
-@Service("nmqsService")
-public class NmqsServiceImpl implements NmqsService {
-    private NmqsMapper nmqsMapper;
-    private MqPageService mqPageService;
-    @Autowired
-    public void setNmqsMapper(NmqsMapper nmqsMapper) {
-        this.nmqsMapper = nmqsMapper;
-    }
-
-    @Autowired
-    public void setMqPageService(MqPageService mqPageService) {
-        this.mqPageService = mqPageService;
-    }
+@Service
+@RequiredArgsConstructor
+public class NmqsServiceImpl extends ServiceImpl<NmqsMapper, NmqsToken> implements NmqsService {
+    private final NmqsMapper nmqsMapper;
+    private final MqPageMapper mqPageMapper;
 
     @Override
     public List<NmqsToken> select(Map<String, Object> params) {
@@ -52,19 +49,12 @@ public class NmqsServiceImpl implements NmqsService {
     @Override
     public int deleteById(Long id) {
         //查询id下的token是否有页面正在使用
-        NmqsToken nmqsToken = nmqsMapper.selectById(id);
-
-        List<MQPage> select = mqPageService.select(null);
-        for (MQPage mqPage : select) {
-            if (mqPage.getToken().equals(nmqsToken.getToken())) {
-                return -1;
-            }
+        QueryWrapper<MQPage> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("nmqs_id", id);
+        List<MQPage> mqPages = mqPageMapper.selectList(queryWrapper);
+        if (!mqPages.isEmpty()) {
+            return -1;
         }
         return nmqsMapper.deleteById(id);
-    }
-
-    @Override
-    public NmqsToken getInfoByToken(String token) {
-        return nmqsMapper.getInfoByToken(token);
     }
 }
